Add new general load operators
These operators take an array and distribute it among threads like
[0, 1, 2, 3, ...] -> t0 -> (0, w, 2w, ...), t1 -> (1, w + 1, ...).
If the number of slots in threads * registers exceeds the capacity of
the input, the load either wraps around (load_rap) or truncates to the
garbage value (load_trunc).

If the input to a load is multi-dimensional, all but the first
dimension gets assumed to be something that needs to be copied over to
the output array unmoved. That is, we need the shapes to have the form
[i, j1, j2, ... jk] and [o1, o2, ..., j1, j2, ... jk], where k tends
to be 0.

We've also swapped trove to have the conventional threads x registers
form, without breaking any of the code.

... come ta think it, having it be iterations x threads might make
more sense, but that can be fixed later.

This commit also introduces the fact that, if the gather index is
outside the bounds of the array, we return \bot instead of panicking.
This allows conditional drops to exist.