Fix a (massive!) perf bug in branching search, update logging
NOTE: the specs are off by one.

The bug was this:
Testing the state that reaches a ronud of search() and then testing all compositions
of operators afterward is equivalent to first composing and then pnuning the result
(except at the edges) ... except when dealing with multiple branches.
Then, the next level might not perform the expected pruning check.

Therefore, we redefine fields in SynthesisLevels as follows
- matrix is the matrix which tests if there's a point in continuing
*after* applying an operator from that level.
This means we no longer support pruning right at the start of the search,
but that feature was marginal anyway.
Similarly, expected_syms is the index of the set of expected symbols
to use after the operator is applied.

Results show *massive* improvements in speed from this change,
and validate the potential fold pruning rule.

This also allowed us to get rid of the consecutive merges restriction
which was weird and and dumb.

The caching code has been fully disabled, and the stats output has been updated.

Finally, we fixed a minor bug in the output relating to spec: lines,
and bumped the ndarray version.